requirements.txt
django
openai
python-dotenv
gunicorn
psycopg2-binary
requests
django-social-share
django-cors-headers
django-allauth
django-ckeditor